Lessons 16–17

+++ TestAQA1/DTO/DapperTestsDTO/ProductDTO.cs
@@ -13,7 +13,7 @@ namespace AutomationTests.DTO.DapperTestsDTO
 
         string description,
 
-        long price,
+        double price,
 
         long stock,
